fix: general refactor

+++ test/api-migration-endpoints/utils/compareData.js
@@ -10,18 +10,19 @@ const { postgresUrl, mongoUrl } = config
 chai.use(deepEqualInAnyOrder)
 const { expect } = chai
 
-export async function compareDataFromBothEnvs ({ endpoint, keysToSkip = {data: [], pages: []} }) {
+export async function compareDataFromBothEnvs ({ endpoint, keysToSkip = {} }) {
   const postgresRes = (await axios.get(postgresUrl + endpoint)).data
   const mongoRes = (await axios.get(mongoUrl + endpoint)).data
+
   const { data: postgresData, pages: postgresPages } = postgresRes
   const { data: mongoData, pages: mongoPages } = mongoRes
 
   try {
-    if (keysToSkip.data.length) {
+    if (keysToSkip.data && keysToSkip.data.length) {
       deleteKeys(postgresData, mongoData, keysToSkip.data)
     }
 
-    if (keysToSkip.pages.length) {
+    if (keysToSkip.pages && keysToSkip.pages.length) {
       deleteKeys(postgresPages, mongoPages, keysToSkip.pages)
     }
 
@@ -30,6 +31,6 @@ export async function compareDataFromBothEnvs ({ endpoint, keysToSkip = {data: [
     const dataDifferences = getObjectDifferences({ postgresObject: postgresData, mongoObject: mongoData })
     const pagesDifferences = getObjectDifferences({ postgresObject: postgresPages, mongoObject: mongoPages })
     console.dir({ dataDifferences, pagesDifferences }, { depth: null })
-    throw new Error(`Results are not equals when testing endpoint ${endpoint}, see differences logged above`)
+    throw new Error(`Results are not equal when testing endpoint ${endpoint}, see differences logged above`)
   }
 }
